requirements.txt
selenium==3.12.0
Pillow==6.1.0
Appium_Python_Client==0.46
PyYAML==5.1.2
